requirements.txt
discord.py-self
pytermx